apps/api/pyproject.toml
[tool.poetry]
name = "vet-researcher-api"
version = "0.1.0"
description = "API service for Vet Researcher"
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "api", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
python-dotenv = "^1.0.0"
pydantic = "^2.6.1"
vet-researcher-core = {path = "../../packages/vet-researcher-core"}

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "api.main:start" 